refactor: 토스트 파일 내부로 context 파일 이동
