+++ src/FunEatProvider.tsx
@@ -1,9 +1,9 @@
 import type { PropsWithChildren } from 'react';
 import { ThemeProvider } from 'styled-components';
 
-import ToastProvider from './contexts/ToastContext';
 import GlobalStyle from './styles/GlobalStyle';
 import theme from './styles/theme';
+import ToastProvider from './Toast/context/ToastContext';
 
 const FunEatProvider = ({ children }: PropsWithChildren) => {
   return (
