+++ src/Toast/context/index.ts
@@ -1,2 +1,3 @@
 export { default as useToastActionContext } from './useToastActionContext';
 export { default as useToastValueContext } from './useToastValueContext';
+export { default as ToastContext } from './ToastContext';
